backend/pyproject.toml
[tool.poetry]
name = "transpocargo"
version = "0.1.0"
description = "TRANSPOCARGO backend code"

authors = [
    "Transpocargo LTD",
]

repository = ""
homepage = ""

# README file(s) are used as the package description
readme = ["README.md", ]

# Keywords (translated to tags on the package index)
keywords = ["transpocargo", "transpocargo", "backend"]

[tool.poetry.dependencies]
# Compatible Python versions
python = ">=3.11,<4"
django = "=5.0.4"
django-environ = "~=0.11.2"
django-storages = "~=1.14.2"
psycopg = {extras = ["binary", "pool"], version = "^3.1.9"}
waitress = "~=3.0.0"
whitenoise = "~=6.6.0"
djangorestframework = "~=3.15.1"
dj-rest-auth= "~=4.0.1"
djangorestframework-simplejwt= "~=5.3.1"
django-extensions = "~=3.2.3"
django-cors-headers = "~=4.3.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}
drf-extra-fields = "~=3.7.0"
django-nested-admin = "~=4.0.2"
drf-writable-nested = "~=0.7.0"
drf-nested-routers = "~=0.93.5"
django-model-utils = "~=4.5.0"
munch = "~=4.0.0"
sentry-sdk = "~=1.45.0"
drf-spectacular = "^0.27.2"
pillow = "~=9.5.0"
django-json-widget = "2.0.1"
django-jazzmin = "^2.6.1"
celery = "^5.3.6"
django-encrypted-model-fields = "^0.6.5"
beautifulsoup4 = "^4.12.3"
boto3 = "^1.34.84"
django-webshell = "^0.4.0"
django-user-agents = "^0.4.0"
django-timezone-field = "^6.0.1"
django-redis = "^5.4.0"
django-cities-light = "^3.9.2"
django-money = "^3.5.2"
reportlab = "^4.2.0"


[tool.chuy]
migrate = "python manage.py migrate"
mig = "python manage.py migrate"
mkmig = "python manage.py makemigrations"
mkmigmerge = "python manage.py makemigrations --merge"
sp = "python manage.py shell_plus"
